WriteLikeMe/requirements.txt
flask==2.0.1
Pillow==9.0.0
python-dotenv==0.19.0 